+++ src/sg_command.h
@@ -480,6 +480,7 @@ struct SG_Command_TextDefaultFont : public SG_Command {
 struct SG_Command_TextRebuild : public SG_Command {
     SG_ID text_id; // lazily create text if not found
     SG_ID material_id;
+    SG_ID geo_id;
     glm::vec2 control_point; // TODO do as material uniform
     float vertical_spacing;
     ptrdiff_t font_path_str_offset;